Move gamepad logic from FetchMessage() to GameEventHandler()
